Modify Auth to have seperate sign in / sign up screens
